requirements.txt
boto3
tqdm
psycopg2-binary
python-dotenv
requests